Bump to version 1.0.0
This version contains many new features.

Note that this version contains breaking changes that make it
incompatible with earlier versions because of database schema changes.
I'm releasing this with the breaking changes because I don't believe
anyone is using this project who is not on the development branch.

New features include:

* Add npm "start" script
* Fix symbile file name for .pdb files
* Add originating request IP
* Add `extensions` map for file extensions of files
* Npm script for running the linter
* Change customFields options to be objects
* Add /api/crashreport api

Thanks to @Jimbly for lots of great contributions.